feat: use MessageBox
