+++ src/modules/situations/index.ts
@@ -1,4 +1,4 @@
 export { SituationOrNoticeIcon } from './situation-or-notice-icon';
-export { SituationMessageBox, MessageBox } from './situation-message-box';
+export { SituationMessageBox } from './situation-message-box';
 export * from './types';
 export * from './utils';
