+++ src/modules/situations/situation-message-box.tsx
@@ -1,23 +1,20 @@
-import { Theme } from '@atb-as/theme';
 import {
   getMessageTypeForSituation,
   getSituationSummary,
   messageTypeToColorIcon,
-  messageTypeToMonoIcon,
   validateEndTime,
 } from './utils';
 import { ColorIcon, MonoIcon } from '@atb/components/icon';
 import { Typo } from '@atb/components/typography';
 import { Situation } from './types';
 import { useTranslation, getTextForLanguage } from '@atb/translations';
-import { useTheme } from '@atb/modules/theme';
 import style from './situation.module.css';
 import { Button } from '@atb/components/button';
 import { useRef } from 'react';
 import dictionary from '@atb/translations/dictionary';
 import { Situation as SituationTexts } from '@atb/translations/modules';
 import { formatToLongDateTime } from '@atb/utils/date';
-import { andIf } from '@atb/utils/css';
+import { MessageBox, MessageBoxProps } from '@atb/components/message-box';
 
 export type Props = {
   situation: Situation;
@@ -86,64 +83,6 @@ export const SituationMessageBox = ({
         onClick={() => dialogRef.current?.showModal()}
       />
     </>
-  );
-};
-
-export type MessageBoxProps = {
-  type: keyof Theme['static']['status'];
-  title?: string;
-  message: string;
-  noStatusIcon?: boolean;
-  onClick?: () => void;
-  borderRadius?: boolean;
-};
-
-export const MessageBox = ({
-  noStatusIcon,
-  type,
-  message,
-  title,
-  onClick,
-  borderRadius = true,
-}: MessageBoxProps) => {
-  const { static: staticColors } = useTheme();
-  const { t } = useTranslation();
-  const backgroundColorStyle = {
-    backgroundColor: staticColors['status'][type].background,
-    color: staticColors['status'][type].text,
-  };
-
-  return (
-    <div
-      className={andIf({
-        [style.container]: true,
-        [style.borderRadius]: borderRadius,
-      })}
-      style={backgroundColorStyle}
-    >
-      {!noStatusIcon && (
-        <MonoIcon className={style.icon} icon={messageTypeToMonoIcon(type)} />
-      )}
-      <div className={style.content}>
-        {title && (
-          <Typo.h2 className={style.title} textType="body__primary--bold">
-            {title}
-          </Typo.h2>
-        )}
-        <Typo.p className={style.body} textType="body__primary">
-          {message}
-        </Typo.p>
-        {onClick && (
-          <Button
-            mode="transparent--underline"
-            onClick={() => onClick()}
-            title={t(dictionary.readMore)}
-            className={style.messageBox__button}
-            size="compact"
-          />
-        )}
-      </div>
-    </div>
   );
 };
