+++ src/page-modules/assistant/layout.tsx
@@ -18,6 +18,7 @@ import { useRouter } from 'next/router';
 import { FormEventHandler, PropsWithChildren, useState } from 'react';
 import style from './assistant.module.css';
 import { createTripQuery } from './utils';
+import { MessageBox } from '@atb/components/message-box';
 
 export type AssistantLayoutProps = PropsWithChildren<{
   initialFromFeature?: GeocoderFeature;
@@ -140,10 +141,7 @@ function AssistantLayout({
           </div>
 
           {geolocationError !== null && (
-            <div className={style.geolocationError}>
-              <MonoIcon icon="status/Info" />
-              <Typo.p textType="body__primary">{geolocationError}</Typo.p>
-            </div>
+            <MessageBox type="warning" message={geolocationError} />
           )}
         </motion.div>
